Move module to common/cs and use it in more places
