+++ background/handle-auth.js
@@ -1,7 +1,7 @@
 import { startCache } from "./message-cache.js";
 import { openMessageCache } from "../libraries/common/message-cache.js";
 import { purgeDatabase } from "../addons/scratch-notifier/notifier.js";
-import { isFirefox } from "../libraries/common/detect-browser.js";
+import { isFirefox } from "../libraries/common/cs/detect-browser.js";
 
 async function getDefaultStoreId() {
   const CHROME_DEFAULT = "0";
