+++ background/imports/util.js
@@ -1,4 +1,4 @@
-import { isFirefox } from "../../libraries/common/detect-browser.js";
+import { isFirefox } from "../../libraries/common/cs/detect-browser.js";
 
 // REMINDER: update similar code at /webpages/settings/index.js
 const browserLevelPermissions = ["notifications"];
