+++ libraries/common/cs/detect-browser.js
@@ -1,0 +1,3 @@
+export function isFirefox() {
+  return typeof new Error().fileName !== "undefined";
+}
