--- libraries/common/detect-browser.js
@@ -1,3 +1,0 @@
-export function isFirefox() {
-  return typeof browser?.runtime.getBrowserInfo === "function";
-}
