+++ libraries/common/notification-util.js
@@ -1,3 +1,5 @@
+import { isFirefox } from "./cs/detect-browser.js";
+
 let id = 0;
 
 export default function create(opts) {
@@ -7,7 +9,7 @@ export default function create(opts) {
   if (scratchAddons.muted) return Promise.resolve(null);
   const notifId = `${opts.base}__${Date.now()}_${id++}`;
   let newOpts;
-  if (!/Chrom/.test(navigator.userAgent)) {
+  if (isFirefox()) {
     newOpts = JSON.parse(JSON.stringify(opts));
     // On Firefox, remove notification properties that throw.
     delete newOpts.buttons;
